+++ CHANGES.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
+#### Changed
+
+  - Added `reg-event-error-handler` to provide a way to handle errors in the interceptor chain [#423](https://github.com/Day8/re-frame/pull/423).
+
 #### Fixed
 
   - `after` and `enrich` interceptors now no longer run if there is no `db` effect, rather than running against the `db` coffect. [#453](https://github.com/Day8/re-frame/issues/453)